Fix Go tests and release verification (#750)
The Go tests had a `replace` statement in them. This breaks the
verify-release action's expectations.

The good news here is that the last three releases weren't actually
broken; but now the release verification does the right thing. Also
added the ability to run the Go tests locally.

A separate run of the verify release job correctly verifies the latest
release and shows this fix is working:
https://github.com/pulumi/pulumi-tls/actions/runs/14063534937/job/39380119347

Fixes #741.

- **Remove replace statements from go.mod**
- **Make sure test can run locally**